Crate ab_farmer_components

Components of the reference implementation of Subspace Farmer for Subspace Network Blockchain.

These components are used to implement farmer itself, but can also be used independently if necessary.

Modules

auditing

Auditing utilities

file_ext

File extension trait

plotting

Plotting utilities

proving

Utilities for turning solution candidates (from auditing) into solutions (proving)

reading

Reading utilities

sector

Sector-related data structures

shard_commitment

Utilities related to shard commitments

Structs

AsyncReadBytes

Container or asynchronously reading bytes using in ReadAtAsync

FarmerProtocolInfo

Information about the protocol necessary for farmer operation

ReadAtOffset

Reader with fixed offset added to all attempted reads

Enums

ReadAt

Enum to encapsulate the selection between ReadAtSync and [ReadAtAsync] variants

Traits

ReadAtAsync

Async version of ReadAt, it is neither Send nor Sync and is supposed to be used with concurrent async combinators

ReadAtSync

Sync version of ReadAt, it is both Send and Sync and is supposed to be used with a thread pool