ab_farmer_components::plotting

Function download_sector

pub async fn download_sector<PG>(
    options: DownloadSectorOptions<'_, PG>,
) -> Result<DownloadedSector, PlottingError>
where
    PG: PieceGetter + Send + Sync,

Download sector for plotting.

This will identify necessary pieces and download them using provided piece getter, after which they can be encoded using encode_sector and written to the plot.